Moving vehicleChecks inside of testData
