+++ src/modules/tests/test_data/test-data.selector.ts
@@ -1,8 +1,10 @@
 
-import { TestData, TestRequirements, ETA, Eco, Manoeuvres } from '@dvsa/mes-test-schema/categories/B';
+import { TestData, TestRequirements, ETA, Eco, Manoeuvres, VehicleChecks } from '@dvsa/mes-test-schema/categories/B';
 import { Competencies, LegalRequirements, ExaminerActions } from './test-data.constants';
 import { pickBy, sumBy, endsWith, get } from 'lodash';
 import { CompetencyOutcome } from '../../../shared/models/competency-outcome';
+import { default as tellMeQuestions } from '../../../providers/question/tell-me-question.constants';
+import { default as showMeQuestions } from '../../../providers/question/show-me-question.constants';
 
 export const getDrivingFaultCount = (data: TestData, competency: Competencies) => data.drivingFaults[competency];
 
@@ -78,3 +80,24 @@ export const hasControlledStopBeenCompleted = (data: TestData) => data.controlle
 export const hasLegalRequirementBeenCompleted = (data: TestRequirements, legalRequirement: LegalRequirements) => {
   return data[legalRequirement];
 };
+
+export const getVehicleChecks = (state: TestData): VehicleChecks => state.vehicleChecks;
+
+export const isTellMeQuestionSelected = (state: VehicleChecks) => get(state, 'tellMeQuestion.code') !== undefined;
+
+export const isTellMeQuestionCorrect = (state: VehicleChecks) => get(state, 'tellMeQuestion.outcome') === 'P';
+
+export const isTellMeQuestionDrivingFault = (state: VehicleChecks) =>
+  get(state, 'tellMeQuestion.outcome') === CompetencyOutcome.DF;
+
+export const getSelectedTellMeQuestionText = (state: VehicleChecks) => {
+  const tellMeQuestionText =
+    tellMeQuestions.find(question => question.code === get(state, 'tellMeQuestion.code'));
+  if (!tellMeQuestionText) {
+    return '';
+  }
+  return `${get(state, 'tellMeQuestion.code')} - ${tellMeQuestionText.shortName}`;
+};
+
+export const getShowMeQuestion = (state: VehicleChecks) =>
+  showMeQuestions.find(question => question.code === get(state, 'showMeQuestion.code'));
